added error classes and auth routes

+++ src/Services/role.service.js
@@ -1,3 +1,4 @@
+const { DuplicateError, NotExistError } = require("../Errors");
 const { RoleRepository } = require("../Repositories");
 
 class RoleService {
@@ -7,6 +8,11 @@ class RoleService {
 
   async create(dataObj) {
     try {
+      const rolExists = await this.roleRepository.findByRoleName(
+        dataObj.roleName
+      );
+      if (rolExists) throw new DuplicateError("role", dataObj.roleName);
+
       const response = await this.roleRepository.create(dataObj);
       return response;
     } catch (error) {
@@ -16,6 +22,9 @@ class RoleService {
 
   async update(id, dataObj) {
     try {
+      const rolExists = await this.roleRepository.findById(id);
+      if (!rolExists) throw new NotExistError("role", dataObj.roleName);
+
       const response = await this.roleRepository.update(id, dataObj);
       return response;
     } catch (error) {
@@ -25,6 +34,9 @@ class RoleService {
 
   async findById(id) {
     try {
+      const rolExists = await this.roleRepository.findById(id);
+      if (!rolExists) throw new NotExistError("roleId", id);
+
       const response = await this.roleRepository.findById(id);
       return response;
     } catch (error) {
@@ -34,6 +46,9 @@ class RoleService {
 
   async delete(id) {
     try {
+      const rolExists = await this.roleRepository.findByRoleName(id);
+      if (!rolExists) throw new NotExistError("role", id);
+
       const response = await this.roleRepository.delete(id);
       return response;
     } catch (error) {
